# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Spotting API
  version: 1.0.0
  description: |
    Spotting REST API (`apps/api/src/v1.ts` + capture public router).
    Authenticated routes use `Authorization: Bearer <session-token>`.
    Org-scoped report routes accept optional `x-org-id`.
  contact:
    name: Spotting Engineering
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://api-spotting.yourdomain.com/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
  parameters:
    OrgIdHeader:
      name: x-org-id
      in: header
      required: false
      schema:
        type: string
...
  schemas:
...
    Report:
      type: object
      properties:
        id:
          type: string
        organizationId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        visibility:
          type: string
          enum: [private, public]
        pageUrl:
          type: string
          nullable: true
        metadataJson:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReportListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
paths:
...
  /reports:
    get:
      tags: [Reports]
      summary: Paginated report list for current org
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrgIdHeader"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Report page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListPage"
    post:
      tags: [Reports]
      summary: Create report (authenticated capture)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrgIdHeader"
      responses:
        "201":
          description: Report created
...
  /reports/upload-sessions:
    post:
      tags: [Reports]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Upload session created
  /reports/upload-sessions/{sessionId}/finalize:
    post:
      tags: [Reports]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Upload finalized
...